# [FEATURE] Printing Optimization of `HemeraTermFx`