--- a/nyx/nyx_engine/utils/nyx_asset_import.py
+++ b/nyx/nyx_engine/utils/nyx_asset_import.py
@@ -21,6 +21,21 @@ class NyxAssetImport:
         """Static method for checking if a file exists before opening to conver the stored array to
         an np.ndarray used in the program."""
         module_path = os.path.dirname(__file__)
-        path = os.path.join(module_path, f"../../../examples/assets/nyx/{file_name}.nyx")
+        path = os.path.join(
+            module_path, f"../../../examples/assets/nyx/{file_name}.nyx"
+        )
         with open(path) as file:
             return np.array(ast.literal_eval(file.read()), dtype=np.uint8)
+
+    @staticmethod
+    def open_npz_asset(file_name: str, file_path: str = None):
+        """Static method for checking if a file exists before opening to conver the stored array to
+        an np.ndarray used in the program."""
+        module_path = os.path.dirname(__file__)
+        path = os.path.join(
+            module_path, f"../../../examples/assets/nyx/{file_name}.npz"
+        )
+        try:
+            return np.load(path)
+        except FileNotFoundError:
+            print(f"File '{path}' not found")