# [FEATURE] Printing Optimization of `HemeraTermFx`

## Continuation of PR #28 

## Overview
Continued profiling has led to a further reduction in string buffer generation time. 

---

## Changes:
- Create dict maps for fg and bg colors, generated at runtime to further limit string manipulation within the loop
- Improve sum check and bg retrieval by splitting the array and ensuring `uint16` datatype for the summed array.
- Moved line profiling into its own wrapper method with logic to activate or bypass.
- Added a method to import npz (multiple ndarrays) from file quickly (used during testing)
- Updated docstrings across the project.

---

### cProfile
Current breakdown for `HemeraTermFx`:
- Printing to terminal: ~25% of total CPU time 
- Generating the String Buffer: ~75% of CPU time

### Line Profiling
- Metrics from `_generate_string_buffer()` method for frames 1-329.
  
  | Version | Total Time | Per Frame Time | FPS |
  |--------|--------|--------|--------|
  | **v0.1.1-alpha** | **4.64321 s** | **0.0141 s** | **70.9** |
  | v0.1.0-alpha | 10.1012 s | 0.0307 s | 32.6 |
  | v0.0.4-alpha | 40.4568 s | 0.1230 s | 8.1 |

---

Line Profile:

```plaintext
Total time: 4.64321 s
File: /home/charles/projects/nyx-engine/nyx/hemera_term_fx/hemera_term_fx.py
Function: _generate_string_buffer at line 134

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   134                                               def _generate_string_buffer(self, delta_frame: np.ndarray):
   135                                                   """Convert the delta frame to its color-formatted `str` representation form before printing
   136                                                   to the terminal.
   137                                                   Args:
   138                                                       delta_frame (np.ndarray): The delta frame to process and print.
   139                                                   """
   140                                                   # Calculate sum of fg + bg
   141       329    7946062.0  24152.2      0.2          delta_frame, sum_frame = self.sum_bg(delta_frame)
   142                                                   # Calculate the sum of each row
   143       329   11608907.0  35285.4      0.3          row_sums = np.sum(sum_frame, axis=1)
   144                                                   # Start the string buffer
   145       329     582424.0   1770.3      0.0          buffer = io.StringIO()
   146                                                   # Initialize loop variables outside the loop
   147       329     587134.0   1784.6      0.0          empty_pixel = np.uint8(0)
   148       329     340558.0   1035.1      0.0          empty_sum = np.uint16(0)
   149       329      60799.0    184.8      0.0          fg_color, bg_color = empty_pixel, empty_sum
   150       329      33923.0    103.1      0.0          last_ansi_fg_color, last_ansi_bg_color = empty_pixel, empty_sum
   151       329      29889.0     90.8      0.0          sum_color, last_subpixel_sum = empty_sum, empty_sum
   152       329     104728.0    318.3      0.0          h, w = delta_frame.shape
   153       329      47790.0    145.3      0.0          row_buffer = []
   154       329      45990.0    139.8      0.0          ansi_fg = self.ansi_fg
   155       329      31493.0     95.7      0.0          ansi_bg = self.ansi_bg
   156                                           
   157                                                   # Iterate frame
   158     29939    3270958.0    109.3      0.1          for y in range(h):
   159     29610    6358300.0    214.7      0.1              row_buffer.clear()
   160                                           
   161                                                       # Check if row has changes
   162     29610    7099390.0    239.8      0.2              if row_sums[y] > empty_sum:
   163   9355931  946538304.0    101.2     20.4                  for x in range(w):
   164   9336480 1385862925.0    148.4     29.8                      sum_color = sum_frame[y, x]
   165                                           
   166                                                               # If the current pixel is printable, get the fg color and calculate the bg color
   167   9336480  999879070.0    107.1     21.5                      if sum_color != empty_sum:
   168    285562   50180100.0    175.7      1.1                          fg_color = delta_frame[y, x]
   169    285562   53726218.0    188.1      1.2                          bg_color = sum_color - fg_color
   170                                                                   # bg_color = delta_frame[1, y, x]
   171                                           
   172                                                                   # Skip cursor movement if it's the same row/column as the last printed pixel
   173    285562   27226889.0     95.3      0.6                          if last_subpixel_sum == empty_sum:
   174     62820   22341331.0    355.6      0.5                              row_buffer.append(f"\033[{y + 1};{x + 1}H")
   175                                           
   176                                                                   # Only write color change sequences when necessary (skip if same as last)
   177                                                                   # Foreground color check/caching
   178    285562   34013346.0    119.1      0.7                          if fg_color != last_ansi_fg_color:
   179    102022   16855453.0    165.2      0.4                              row_buffer.append(ansi_fg[fg_color])
   180    102022   10483737.0    102.8      0.2                              last_ansi_fg_color = fg_color
   181                                                                   # Background color check/caching
   182    285562   31466926.0    110.2      0.7                          if bg_color != last_ansi_bg_color:
   183     80863   13172782.0    162.9      0.3                              row_buffer.append(ansi_bg[bg_color])
   184     80863    7909458.0     97.8      0.2                              last_ansi_bg_color = bg_color
   185                                           
   186                                                                   # Add the printed character
   187    285562   31807907.0    111.4      0.7                          row_buffer.append("▀")
   188                                           
   189                                                               # Cache the last sum
   190   9336480  930544325.0     99.7     20.0                      last_subpixel_sum = sum_color
   191                                           
   192                                                           # Add the row buffer for the changed row
   193     19451   17527523.0    901.1      0.4                  buffer.write("".join(row_buffer) + "\n")
   194                                           
   195                                                   # Output the accumulated buffer to stdout
   196       329   24500136.0  74468.5      0.5          self.write_to_term(buffer.getvalue())
   197       329    1026467.0   3120.0      0.0          self.flush_to_term()
```